refactor(MCP): relocate constants to `server_constants` for improved modularity
