+++ front/lib/actions/mcp_internal_actions/servers/agent_memory.ts
@@ -4,7 +4,7 @@ import assert from "assert";
 import { z } from "zod";
 
 import type { MCPError } from "@app/lib/actions/mcp_errors";
-import { AGENT_MEMORY_SERVER_NAME } from "@app/lib/actions/mcp_internal_actions/constants";
+import { AGENT_MEMORY_SERVER_NAME } from "@app/lib/actions/mcp_internal_actions/server_constants";
 import { makeInternalMCPServer } from "@app/lib/actions/mcp_internal_actions/utils";
 import { withToolLogging } from "@app/lib/actions/mcp_internal_actions/wrappers";
 import type { AgentLoopContextType } from "@app/lib/actions/types";
