+++ front/lib/actions/mcp_internal_actions/utils.ts
@@ -6,17 +6,17 @@ import {
 } from "@dust-tt/client";
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 
-import type { InternalMCPServerNameType } from "@app/lib/actions/mcp_internal_actions/constants";
-import {
-  AGENT_MEMORY_SERVER_NAME,
-  INTERNAL_MCP_SERVERS,
-  isInternalMCPServerOfName,
-} from "@app/lib/actions/mcp_internal_actions/constants";
 import type {
   ToolEarlyExitEvent,
   ToolPersonalAuthRequiredEvent,
 } from "@app/lib/actions/mcp_internal_actions/events";
 import { getMCPServerRequirements } from "@app/lib/actions/mcp_internal_actions/input_configuration";
+import type { InternalMCPServerNameType } from "@app/lib/actions/mcp_internal_actions/server_constants";
+import {
+  AGENT_MEMORY_SERVER_NAME,
+  INTERNAL_MCP_SERVERS,
+  isInternalMCPServerOfName,
+} from "@app/lib/actions/mcp_internal_actions/server_constants";
 import type { MCPServerViewType } from "@app/lib/api/mcp";
 import type { AgentMCPActionOutputItem } from "@app/lib/models/assistant/actions/mcp";
 import type { AgentMCPActionResource } from "@app/lib/resources/agent_mcp_action_resource";
