+++ front/lib/agent_yaml_converter/converter.ts
@@ -5,7 +5,7 @@ import { processAdditionalConfiguration } from "@app/components/agent_builder/su
 import {
   isAutoInternalMCPServerName,
   isInternalMCPServerName,
-} from "@app/lib/actions/mcp_internal_actions/constants";
+} from "@app/lib/actions/mcp_internal_actions/server_constants";
 import type { Authenticator } from "@app/lib/auth";
 import { MCPServerViewResource } from "@app/lib/resources/mcp_server_view_resource";
 import { concurrentExecutor } from "@app/lib/utils/async_utils";
