+++ front/lib/api/assistant/global_agents/configurations/dust/deep-dive.ts
@@ -1,6 +1,6 @@
 import { DEFAULT_WEBSEARCH_ACTION_DESCRIPTION } from "@app/lib/actions/constants";
 import type { MCPServerConfigurationType } from "@app/lib/actions/mcp";
-import type { InternalMCPServerNameType } from "@app/lib/actions/mcp_internal_actions/constants";
+import type { InternalMCPServerNameType } from "@app/lib/actions/mcp_internal_actions/server_constants";
 import {
   DEEP_DIVE_AVATAR_URL,
   DEEP_DIVE_DESC,
