+++ front/lib/resources/internal_mcp_server_in_memory_resource.ts
@@ -5,7 +5,7 @@ import {
   internalMCPServerNameToSId,
 } from "@app/lib/actions/mcp_helper";
 import { isEnabledForWorkspace } from "@app/lib/actions/mcp_internal_actions";
-import type { InternalMCPServerNameType } from "@app/lib/actions/mcp_internal_actions/constants";
+import type { InternalMCPServerNameType } from "@app/lib/actions/mcp_internal_actions/server_constants";
 import {
   allowsMultipleInstancesOfInternalMCPServerById,
   allowsMultipleInstancesOfInternalMCPServerByName,
@@ -16,7 +16,7 @@ import {
   isAutoInternalMCPServerName,
   isInternalMCPServerName,
   isInternalMCPServerOfName,
-} from "@app/lib/actions/mcp_internal_actions/constants";
+} from "@app/lib/actions/mcp_internal_actions/server_constants";
 import {
   connectToMCPServer,
   extractMetadataFromServerVersion,
